rust: add rust specific type inference
Fixes compile errors and some semantic errors. More work needed
for cases like:

```
fn add1(x: i8, y: i8) -> i16 {
-    return (x + y) as i16;
+    return (x as i16 + y as i16);
 }
```

so add1(127, 10) returns 137 instead of -119.

Related to https://github.com/adsharma/py2many/issues/16

+++ tests/expected/binit.rs
@@ -5,10 +5,10 @@ fn bisect_right(data: &Vec<i32>, item: i32) -> i32 {
     let mut high: i32 = data.len() as i32;
     while low < high {
         let middle: _ = i32::from(((low + high) / 2));
-        if item < data[middle as usize] {
+        if item < data[middle as usize] as i32 {
             high = middle;
         } else {
-            low = (middle + 1);
+            low = (middle as i32 + 1);
         }
     }
     return low;
